requirements.txt
Flask==3.0.3
pandas==2.2.2
PyYAML==6.0.2
python-dateutil==2.9.0.post0
requests==2.32.3